remove login service
