--- backend/services/interface/loginInterface.ts
@@ -1,12 +1,0 @@
-interface ILoginService {
-  adminLogin(
-    role: string,
-    password: string,
-  ): Promise<{ token: string }>;
-  participantLogin(
-    id: number,
-    password: string,
-  ): Promise<{ token: string }>;
-}
-
-export default ILoginService;
